Edit: Fix document expansion for doc
